+++ src/main/java/com/infeco/keylease/exceptions/NotFoundEntity.java
@@ -1,0 +1,4 @@
+package com.infeco.keylease.exceptions;
+
+public class NotFoundEntity extends Exception {
+}