+++ src/main/java/com/infeco/keylease/models/Address.java
@@ -2,10 +2,7 @@ package com.infeco.keylease.models;
 
 import com.infeco.keylease.entity.AddressEntity;
 
-import java.util.UUID;
-
 public class Address {
-    UUID id;
     String street;
     String additionalAddress;
     String zipCode;
@@ -15,18 +12,9 @@ public class Address {
     }
 
     public Address(AddressEntity addressEntity) {
-        this.id = addressEntity.getId();
         this.street = addressEntity.getStreet();
         this.additionalAddress = addressEntity.getAdditionalAddress();
         this.zipCode = addressEntity.getZipCode();
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
     }
 
     public String getStreet() {