+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Nomenclature :
 
 ### Unreleased
 
+### Added
+
+- PUT /tenants with Bad Request Exception
+
 ### [0.0.6] - 2023-02-22
 
 #### Added